refactor(notifications): extract validation interface to validation.rs
Move BlockChecked trait, ValidationInterfaceCallbacks struct, and
validation wrapper functions to dedicated validation.rs module.
