+++ src/notifications/mod.rs
@@ -1,9 +1,12 @@
 pub mod notification;
 pub mod types;
+pub mod validation;
+
+pub use types::{BlockValidationResult, SynchronizationState, ValidationMode, Warning};
 
 pub use notification::{
     BlockTip, FatalError, FlushError, HeaderTip, KernelNotificationInterfaceCallbacks, Progress,
     WarningSet, WarningUnset,
 };
 
-pub use types::{BlockValidationResult, SynchronizationState, ValidationMode, Warning};
+pub use validation::{BlockChecked, ValidationInterfaceCallbacks};
